move spinner to top
